+++ internal/extsvc/github/user.go
@@ -9,13 +9,13 @@ import (
 	"golang.org/x/oauth2"
 )
 
-func GetExternalAccountData(data *extsvc.ExternalAccountData) (usr *github.User, tok *oauth2.Token, err error) {
+func GetExternalAccountData(data *extsvc.AccountData) (usr *github.User, tok *oauth2.Token, err error) {
 	var (
 		u github.User
 		t oauth2.Token
 	)
 
-	if data.AccountData != nil {
+	if data.Data != nil {
 		if err := data.GetAccountData(&u); err != nil {
 			return nil, nil, err
 		}
@@ -30,7 +30,7 @@ func GetExternalAccountData(data *extsvc.ExternalAccountData) (usr *github.User,
 	return usr, tok, nil
 }
 
-func SetExternalAccountData(data *extsvc.ExternalAccountData, user *github.User, token *oauth2.Token) {
+func SetExternalAccountData(data *extsvc.AccountData, user *github.User, token *oauth2.Token) {
 	data.SetAccountData(user)
 	data.SetAuthData(token)
 }