+++ migrations/1528395667_index_boolean_fields_on_repo.up.sql
@@ -1,0 +1,7 @@
+BEGIN;
+
+CREATE INDEX IF NOT EXISTS repo_private ON repo(private);
+CREATE INDEX IF NOT EXISTS repo_fork ON repo(fork);
+CREATE INDEX IF NOT EXISTS repo_archived ON repo(archived);
+
+COMMIT;